Add PhantomJS to deps and freeze dependencies
It seems there are some issues locally and in CI caused by Ember not
finding PhantomJS in the system, which are easily solved by just
adding it to the project dependencies and installing with NPM.

This commit also introduces npm-shrinkwrap.json to freeze the whole
dependencies tree. We should keep updating that file so that when
something breaks we can revert. Otherwise, things might silently
break due to nested dependencies that don't have the version fixed
in the package.json. More info: https://docs.npmjs.com/cli/shrinkwrap